requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2
django-filter==2.1.0
django-mysql==3.0.0.post1
django-rest-auth==0.9.5
django-rest-authtoken==1.2.2
djangorestframework==3.9.2
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
Pillow==6.0.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
wrapt==1.11.1
